+++ src/stores/data.ts
@@ -1,4 +1,4 @@
-import { writable, readable, derived } from 'svelte/store'
+import { writable } from 'svelte/store'
 import type {
   MsSinceUnixEpoch,
   Track,